Allowed for adjustment of emission rates for alpha sources and handling of decimal emission rates

+++ Rutherford Scattering/src/RutherfordScattering/AlphaSource.cpp
@@ -75,12 +75,12 @@ void RutherfordScattering::AlphaSource::IncrementFrame()
 
 }
 
-void RutherfordScattering::AlphaSource::SetEmissionRate(int newEmissionRate)
+void RutherfordScattering::AlphaSource::SetEmissionRate(float newEmissionRate)
 {
 	_emissionRate = newEmissionRate;
 }
 
-int RutherfordScattering::AlphaSource::GetEmissionRate()
+float RutherfordScattering::AlphaSource::GetEmissionRate()
 {
 	return _emissionRate;
 }
